Improve error message for invalid privkey length
This adds detail to the error message when a Uint8Array is provided as
a privkey but it does not have the correct length. Previously, the error
message stated only "Uint8Array expected" - which is confusing because a
Uint8Array was in fact provided. This should make debugging invalid
privkeys easier.

A test case is added that checks for the new error message.